% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr-package.R
\docType{package}
\name{rdr-package}
\alias{rdr}
\alias{rdr-package}
\title{rdr: Personal Package for Russell Dinnage}
\description{
Random useful stuff for me!
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
